backend/requirements.txt
flask
flask-cors
torch
transformers
nltk
python-docx
PyPDF2